+++ src/lib/components/App/App.js
@@ -11,7 +11,7 @@ import Routing from '../Routing';
 import Header from '../Header';
 import Footer from '../Footer';
 
-import '../../style/core/css/cagov.core.css';
+import '../../style/core/css/cagov.core.min.css';
 import '../../style/style.scss';
 
 /**