+++ src/lib/components/FieldCheckboxes/Checkbox.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import T from 'prop-types';
+import classNames from 'classnames';
 
 import * as propUtils from '../../utils/propTypes';
 
@@ -17,40 +18,59 @@ export const handleChange = ({ value, values, onChange }) => (event) => {
     // not checked, remove from array
     newValues.splice(newValues.indexOf(value), 1);
   }
-  // update in redux form
+
   return onChange(newValues);
 };
 
 // individual checkbox, applies state-template styling
 const Checkbox = (props) => {
   const {
-    name,
-    value,
-    option,
-    variant,
-    disabled,
-    onChange,
-    'aria-invalid': ariaInvalid,
     'aria-describedby': ariaDescribedby,
+    'aria-invalid': ariaInvalid,
+    disabled,
+    inline,
+    name,
+    onBlur,
+    onChange,
+    onFocus,
+    option,
+    value,
+    variant,
   } = props;
 
   const values = Array.isArray(value) ? value : [value];
 
   const id = `${name}-${option.value}`;
   const inputProps = {
+    'aria-describedby': ariaDescribedby,
+    'aria-invalid': ariaInvalid,
+    checked: values.includes(option.value),
+    disabled,
     id,
     name,
-    disabled,
     type: 'checkbox',
-    checked: values.includes(option.value),
-    onChange: handleChange({ value: option.value, values, onChange }),
-    'aria-invalid': ariaInvalid,
-    'aria-describedby': ariaDescribedby,
   };
+
+  if (onBlur) {
+    inputProps.onBlur = () => onBlur(value);
+  }
+
+  if (onChange) {
+    inputProps.onChange = handleChange({ value: option.value, values, onChange });
+  }
+
+  if (onFocus) {
+    inputProps.onFocus = () => onFocus(value);
+  }
+
+  let cn = classNames([
+    'check',
+    option.className,
+  ]);
 
   if (variant) {
     return (
-      <label data-test={'field--checkbox'} className={'check'} htmlFor={id}>
+      <label data-test={'field__checkbox'} className={cn} htmlFor={id}>
         <input className={'hidden-up pos-abs top-0 left-0'} {...inputProps} />
         <span className={`btn btn-md btn-block color-white--checked bg-${variant}--checked rounded-0`}>
           {option.label}
@@ -59,8 +79,14 @@ const Checkbox = (props) => {
     );
   }
 
+  cn = classNames([
+    'p-l-md m-l-0 m-r-md',
+    option.className,
+    inline ? 'form-check-inline' : 'form-check',
+  ]);
+
   return (
-    <label data-test={'field--checkbox'} className={'form-check-inline p-l-md m-l-0 m-r-md'} htmlFor={id}>
+    <label data-test={'field__checkbox'} className={cn} htmlFor={id}>
       <input className={'hidden-up pos-abs'} {...inputProps} />
       <div className={'check-icon-checkbox'} aria-hidden>
         <i className={'ca-gov-icon-check-mark'} />
@@ -71,8 +97,32 @@ const Checkbox = (props) => {
 };
 
 Checkbox.propTypes = {
+  /** Accessible indicator of related information */
+  'aria-describedby': T.string,
+
+  /** Accessible indicator for errors existing */
+  'aria-invalid': T.string,
+
+  /** Disable the input */
+  disabled: T.bool,
+
+  /** Display inline */
+  inline: T.bool,
+
   /** Name of field */
   name: T.string.isRequired,
+
+  /** Called when checkbox is blurred */
+  onBlur: T.func,
+
+  /** Called when checkbox is changed */
+  onChange: T.func,
+
+  /** Called when checkbox is focused */
+  onFocus: T.func,
+
+  /** Option to select */
+  option: propUtils.option.isRequired,
 
   /** Value of checkbox */
   value: T.oneOfType([
@@ -80,37 +130,21 @@ Checkbox.propTypes = {
     T.array,
   ]).isRequired,
 
-  /** Option to select */
-  option: propUtils.option.isRequired,
-
   /** Use style variant */
   variant: T.oneOf([
-    '',
     'primary',
   ]),
-
-  /** Disable the input */
-  disabled: T.bool,
-
-  /** Label to render for this checkbox */
-  label: T.string,
-
-  /** Called when checkbox changes */
-  onChange: T.func.isRequired,
-
-  /** Accessible indicator for errors existing */
-  'aria-invalid': T.string,
-
-  /** Accessible indicator of related information */
-  'aria-describedby': T.string,
 };
 
 Checkbox.defaultProps = {
-  variant: '',
-  disabled: false,
-  label: null,
-  'aria-invalid': 'false',
   'aria-describedby': null,
+  'aria-invalid': 'false',
+  disabled: false,
+  inline: false,
+  onBlur: null,
+  onChange: null,
+  onFocus: null,
+  variant: null,
 };
 
 export default Checkbox;